Changes to get private-hadoop-common build on branch-2.5.1-mapr
Changes to get private-hadoop-common build on branch-2.5.1-mapr
* Includes pom file version changes. The version is now 2.5.1-mapr-1410
* Apache hadoop common has now added code to RM UI to display cpu cores
* Other minor changes

Testing done
Build private-hadoop-common on 2.5.1-mapr-1410

+++ hadoop-common-project/hadoop-common/src/main/java/org/apache/hadoop/fs/FileSystem.java
@@ -2844,9 +2844,9 @@ public abstract class FileSystem extends Configured implements Closeable {
     public static class StatisticsData {
       volatile long bytesRead;
       volatile long bytesWritten;
-      volatile long readOps;
-      volatile long largeReadOps;
-      volatile long writeOps;
+      volatile int readOps;
+      volatile int largeReadOps;
+      volatile int writeOps;
       /**
        * Stores a weak reference to the thread owning this StatisticsData.
        * This allows us to remove StatisticsData objects that pertain to
